test: update test examples
Adds test coverage for existing logic in `update_quality`, and a pending
test for the new "Conjured" item rules. This helps to limit the scope of
the challenge to refactoring the existing `update_quality` logic and
testing/implementing the new feature, as opposed to adding coverage for
regressions (which is still a good thing to practice, but makes the scope
of the challenge much larger).